+++ src/main/html.ts
@@ -1,26 +1,4 @@
-export * from './createSaxParser';
-export * from './createForgivingSaxParser';
-export * from './createDomParser';
-export * from './createXmlDomParser';
-export * from './createEntitiesDecoder';
+export * from './index';
 export * from './createHtmlSaxParser';
 export * from './createHtmlDomParser';
 export * from './createFromHtmlCharName';
-export {IAttrToken} from './token-types';
-export {IDataToken} from './token-types';
-export {IStartTagToken} from './token-types';
-export {ITagToken} from './token-types';
-export {IToken} from './token-types';
-export {ISaxParser} from './sax-parser-types';
-export {ISaxParserOptions} from './sax-parser-types';
-export {ISaxParserCallbacks} from './sax-parser-types';
-export {ISaxParserDialectOptions} from './sax-parser-types';
-export {DataTokenCallback} from './sax-parser-types';
-export {IForgivingSaxParserOptions} from './sax-parser-types';
-export {IForgivingSaxParserDialectOptions} from './sax-parser-types';
-export {IDomParser} from './dom-parser-types';
-export {IDomParserOptions} from './dom-parser-types';
-export {IDomParserFactoryCallbacks} from './dom-parser-types';
-export {DataNodeFactory} from './dom-parser-types';
-export {IDomParserDialectOptions} from './dom-parser-types';
-export {ICustomSaxParserOptions} from './dom-parser-types';
