+++ src/test/createDomParser.test.ts
@@ -47,7 +47,7 @@ describe('createDomParser', () => {
       createElement: domParserOptions.createElement,
       appendChild: domParserOptions.appendChild,
       saxParserFactory: saxParserFactoryMock,
-      foo: 'bar'
+      foo: 'bar',
     };
 
     createDomParser(options).parse('<a></a>');
