+++ src/test/createXmlDomParser.test.ts
@@ -1,11 +1,5 @@
-import {
-  createXmlDomParser,
-  DomNodeType,
-  IDomAttributeMap,
-  IDomElement,
-  IDomNode,
-  IDomText,
-} from '../main/createXmlDomParser';
+import {createXmlDomParser} from '../main/createXmlDomParser';
+import {DomNodeType, IDomAttributeMap, IDomElement, IDomNode, IDomText} from '../main/dom-types';
 
 export function el(tagName: string, start: number, end: number, selfClosing = false, attrs: IDomAttributeMap = {}, children: Array<IDomNode> = []): IDomElement {
